Rework browser window initialization and app lazy loading
* Move StudioWindow creation to the start of app "ready" to get a window displayed asap
* Split Root out of App.tsx into Workspace.tsx
* Load BuiltinPanelCatalogProvider and Workspace via React.lazy to reduce bundle size.

The reduced bundle size allows for minimize optimization for the prod build to succeed
without node.js heap out-of-memory errors. Minimize on production bundles further reduces the initial script parsing time for v8.
